Close "Received Files" modal if an error occurs
For #13
